Changed Extructure, implemeted postcss modules and cleaned css files
