+++ package.json
@@ -15,6 +15,7 @@
     "graphql": "^14.6.0",
     "graphql-tag": "^2.10.3",
     "node-fetch": "^2.6.0",
+    "postcss-modules": "^1.5.0",
     "react": "^16.12.0",
     "react-apollo": "^3.1.3",
     "react-dom": "^16.12.0",
@@ -45,7 +46,6 @@
     ]
   },
   "devDependencies": {
-    "babel-eslint": "^10.0.3",
-    "postcss-modules": "^1.5.0"
+    "babel-eslint": "^10.0.3"
   }
 }
