+++ src/Components/SignUp/SignUp.js
@@ -1,0 +1,149 @@
+import React, { useState } from 'react'
+import { Link, useHistory } from 'react-router-dom'
+import { useMutation } from '@apollo/react-hooks'
+import { Formik } from 'formik'
+import * as Yup from 'yup'
+import form from '../form.module.css'
+import gql from 'graphql-tag'
+
+const validation = Yup.object().shape({
+  firstName: Yup.string()
+    .max(12)
+    .required('Must enter name'),
+  lastName: Yup.string()
+    .max(12)
+    .required('Must enter name'),
+  username: Yup.string()
+    .max(10, 'Must be less than 10 caracters')
+    .required('Enter your username'),
+  password: Yup.string()
+    .required('Must enter a password')
+    .min(5, 'Must be atleast longer than 5 characters')
+})
+
+const SIGN_UP = gql`
+  mutation sign_up($data: SignupInput!){
+    signup(data: $data){
+      user{
+        id
+        username
+      }
+      jwt
+    }
+  }
+`
+
+function SignUp () {
+  const history = useHistory()
+  const [error, setError] = useState(null)
+  const [signup] = useMutation(SIGN_UP, {
+    onCompleted: RegistrationSuccess,
+    onError: RegistrationFailure
+  }
+  )
+
+  function RegistrationFailure () {
+    return (
+      setError(true)
+    )
+  }
+
+  function RegistrationSuccess () {
+    setError(false)
+    return (
+      setTimeout(Redirect, 1500)
+    )
+  }
+
+  function Redirect () {
+    history.push('/auth/signin')
+  }
+
+  return (
+    <div className={form.outer}>
+      <div className={form.middle}>
+        <div className={form.inner}>
+          <div id={form.formContent}>
+            <h1>Member Registration</h1>
+            {error &&
+              <h3 className={form.error}>  oh snap! Something went wrong  </h3>}
+            {error === false &&
+              <h3 className={form.correct}>  User created successfully  </h3>}
+            <Formik
+              initialValues={{ firstName: '', lastName: '', username: '', password: '' }}
+              validationSchema={validation}
+              onSubmit={async (values, { setSubmitting, resetForm }) => {
+                setSubmitting(true)
+                await signup({
+                  variables: {
+                    data: values
+                  }
+                })
+                setSubmitting(false)
+              }}
+            >
+              {({
+                values,
+                handleChange,
+                handleBlur,
+                handleSubmit,
+                isSubmitting
+              }) => (
+                <form onSubmit={handleSubmit}>
+                  <input
+                    type='text'
+                    className={form.form}
+                    id='firstName'
+                    placeholder='FirstName'
+                    onChange={handleChange}
+                    onBlur={handleBlur}
+                    value={values.firstName}
+                  />
+                  <input
+                    type='text'
+                    className={form.form}
+                    id='lastName'
+                    placeholder='LastName'
+                    onChange={handleChange}
+                    onBlur={handleBlur}
+                    value={values.lastName}
+                  />
+                  <input
+                    type='text'
+                    className={form.form}
+                    id='username'
+                    placeholder='Username'
+                    onChange={handleChange}
+                    onBlur={handleBlur}
+                    value={values.username}
+                  />
+                  <input
+                    type='password'
+                    className={form.form}
+                    id='password'
+                    placeholder='Password'
+                    onChange={handleChange}
+                    onBlur={handleBlur}
+                    value={values.password}
+                  />
+                  <button
+                    type='submit'
+                    className={form.button}
+                    disabled={isSubmitting}
+                  >
+                      Sign up
+                  </button>
+                </form>
+              )}
+            </Formik>
+            <Link to='/auth/signin'>
+                  Do you already have a account?
+            </Link>
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
+
+export default SignUp
