+++ src/index.js
@@ -8,17 +8,19 @@ import { ApolloClient } from 'apollo-client'
 import { ApolloProvider } from 'react-apollo'
 import { InMemoryCache } from 'apollo-cache-inmemory'
 import { HttpLink } from 'apollo-link-http'
-/* import { CachePersistor } from 'apollo-cache-persist'; */
 import * as Storage from './Components/Storage'
+import { CachePersistor } from 'apollo-cache-persist'
 
 const cache = new InMemoryCache()
+
+export const persistor = new CachePersistor({ cache, storage: window.localStorage })
 
 const link = new HttpLink({
   uri: 'http://localhost:3001/graphql',
   headers: { Authorization: `Bearer ${Storage.getToken()}` }
 })
 
-const client = new ApolloClient({
+export const client = new ApolloClient({
   cache,
   link
 })
